FIXUP: [Backport] Security bug 905509
Fix some typos and some replacements that were missing or overlooked
whe backporting due to upstream refactoring.

+++ chromium/cc/paint/paint_typeface_transfer_cache_entry.cc
@@ -62,7 +62,7 @@ uint32_t ClientPaintTypefaceTransferCacheEntry::Id() const {
   return typeface_.sk_id();
 }
 
-size_t ClientPaintTypefaceTransferCacheEntry::SerializedSize() const {
+uint32_t ClientPaintTypefaceTransferCacheEntry::SerializedSize() const {
   return size_;
 }
