+++ chromium/cc/paint/path_transfer_cache_entry.h
@@ -18,12 +18,12 @@ class CC_PAINT_EXPORT ClientPathTransferCacheEntry
   explicit ClientPathTransferCacheEntry(const SkPath& path);
   ~ClientPathTransferCacheEntry() final;
   uint32_t Id() const final;
-  size_t SerializedSize() const final;
+  uint32_t SerializedSize() const final;
   bool Serialize(base::span<uint8_t> data) const final;
 
  private:
   SkPath path_;
-  size_t size_ = 0u;
+  uint32_t size_ = 0u;
 };
 
 class CC_PAINT_EXPORT ServicePathTransferCacheEntry
